Use separate sqlite database for testing
