+++ manual_create_db.py
@@ -1,3 +1,11 @@
+from flask import Flask
 
+import settings
 from models import db
-db.create_all()
+
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://{}:{}@localhost/{}'.format(
+    settings.USER, settings.PASSWD, settings.DATABASE)
+db.init_app(app)
+with app.app_context():
+    db.create_all()
