+++ models.py
@@ -1,17 +1,10 @@
-from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
-import settings
 import json
 
-app = Flask(__name__)
 
-app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://{}:{}@localhost/{}'.format(
-    settings.USER, settings.PASSWD, settings.DATABASE)
+db = SQLAlchemy()
 
-#app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-
-db = SQLAlchemy(app)
 
 NODE_KEY_LIMIT = 42
