+++ requirements.txt
@@ -1,3 +1,4 @@
 flask_sqlalchemy==2.1
+Flask-Testing==0.7.1
 PyMySQL
 Flask-SSLify
