+++ tests/test_case.py
@@ -1,0 +1,17 @@
+from flask_testing import TestCase
+
+from app import app, db
+
+
+class TestCase(TestCase):
+    def create_app(self):
+        app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:////tmp/test.db'
+        db.init_app(app)
+        return app
+
+    def setUp(self):
+        db.create_all()
+
+    def tearDown(self):
+        db.session.remove()
+        db.drop_all()
